A more reliable way to tell if a cluster is ready
You can't get runtime info with a recursive read-resource call, so we
used to combine them and count the entries in the :servers entry to tell
whether the cluster nodes were fully started. But this proved unreliable
with the latest WF9. Instead, we no longer rely on the :servers entry.
Instead, we query the config (no runtime data in the recursive request
for the host controller) to see which servers should auto-start, and
then check that the server-state for each server is "running".

Eventually, we could get the server-group from the auto-start'ed servers
instead of just grabbing the first one, but we'll cross that bridge
another time.